+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 - remove Python 2 support
+- add Model.getParams that returns a dict mapping all parameter names to their values
+- add Model.setParams to set multiple parameters at once using a dict
 
 ## 2.2.3 - 2019-12-10
 ### Added